# UAS/src/app/Models/Laporan.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;

class Laporan extends Model
{
    use HasFactory;

    protected $fillable = [
        'judul',
        'description',
        'kategori',
        'lokasi',
        'foto',
        'status',
    ];
}